+++ packages/core-ethereum/crates/core-ethereum-misc/src/network_registry.rs
@@ -21,8 +21,7 @@ where
 mod tests {
     use core_ethereum_db::{db::CoreEthereumDb, traits::HoprCoreEthereumDbActions};
     use hex_literal::hex;
-    use std::sync::{Arc, Mutex};
-    use utils_db::{db::DB, leveldb::rusty::RustyLevelDbShim};
+    use utils_db::{db::DB, rusty::RustyLevelDbShim};
     use utils_types::{
         primitives::{Address, Snapshot},
         traits::BinarySerializable,
@@ -33,11 +32,8 @@ mod tests {
     const TEST_ACCOUNT: [u8; Address::SIZE] = hex!("3a585656b8bbb14e8aebf89256ce4511fa35ac33");
 
     fn create_mock_db() -> CoreEthereumDb<RustyLevelDbShim> {
-        let opt = rusty_leveldb::in_memory();
-        let db = rusty_leveldb::DB::open("test", opt).unwrap();
-
         CoreEthereumDb::new(
-            DB::new(RustyLevelDbShim::new(Arc::new(Mutex::new(db)))),
+            DB::new(RustyLevelDbShim::new_in_memory()),
             Address::from_bytes(&ADDR).unwrap(),
         )
     }
@@ -52,13 +48,9 @@ mod tests {
             .await
             .unwrap();
 
-        db.add_to_network_registry(
-            &Address::from_bytes(&TEST_ADDR).unwrap(),
-            &Address::from_bytes(&TEST_ACCOUNT).unwrap(),
-            &Snapshot::default(),
-        )
-        .await
-        .unwrap();
+        db.set_allowed_to_access_network(&Address::from_bytes(&TEST_ACCOUNT).unwrap(), true, &Snapshot::default())
+            .await
+            .unwrap();
 
         let is_allowed = super::is_allowed_to_access_network(&db, &Address::from_bytes(&TEST_ACCOUNT).unwrap()).await;