+++ packages/core/crates/core-packet/src/validation.rs
@@ -111,14 +111,12 @@ mod tests {
     use core_types::{
         account::AccountEntry,
         channels::{ChannelEntry, Ticket},
-        protocol::TagBloomFilter,
     };
     use hex_literal::hex;
     use lazy_static::lazy_static;
     use mockall::mock;
-    use std::sync::{Arc, Mutex};
     use utils_db::db::DB;
-    use utils_db::leveldb::rusty::RustyLevelDbShim;
+    use utils_db::rusty::RustyLevelDbShim;
     use utils_types::primitives::{
         Address, AuthorizationToken, Balance, BalanceType, EthereumChallenge, Snapshot, U256,
     };
@@ -213,18 +211,6 @@ mod tests {
             async fn sub_hopr_balance(&mut self, balance: &Balance, snapshot: &Snapshot) -> core_ethereum_db::errors::Result<()>;
             async fn is_network_registry_enabled(&self) -> core_ethereum_db::errors::Result<bool>;
             async fn set_network_registry(&mut self, enabled: bool, snapshot: &Snapshot) -> core_ethereum_db::errors::Result<()>;
-            async fn add_to_network_registry(
-                &mut self,
-                public_key: &Address,
-                account: &Address,
-                snapshot: &Snapshot,
-            ) -> core_ethereum_db::errors::Result<()>;
-            async fn remove_from_network_registry(
-                &mut self,
-                public_key: &Address,
-                account: &Address,
-                snapshot: &Snapshot,
-            ) -> core_ethereum_db::errors::Result<()>;
             async fn is_eligible(&self, account: &Address) -> core_ethereum_db::errors::Result<bool>;
             async fn store_authorization(&mut self, token: AuthorizationToken) -> core_ethereum_db::errors::Result<()>;
             async fn retrieve_authorization(&self, id: String) -> core_ethereum_db::errors::Result<Option<AuthorizationToken>>;
@@ -570,11 +556,8 @@ mod tests {
 
     #[async_std::test]
     async fn test_ticket_workflow() {
-        let level_db = Arc::new(Mutex::new(
-            rusty_leveldb::DB::open("test", rusty_leveldb::in_memory()).unwrap(),
-        ));
         let mut db = CoreEthereumDb::new(
-            DB::new(RustyLevelDbShim::new(level_db)),
+            DB::new(RustyLevelDbShim::new_in_memory()),
             SENDER_PRIV_KEY.public().to_address(),
         );
 
@@ -616,11 +599,8 @@ mod tests {
 
     #[async_std::test]
     async fn test_db_should_store_ticket_index() {
-        let level_db = Arc::new(Mutex::new(
-            rusty_leveldb::DB::open("test", rusty_leveldb::in_memory()).unwrap(),
-        ));
         let mut db = CoreEthereumDb::new(
-            DB::new(RustyLevelDbShim::new(level_db)),
+            DB::new(RustyLevelDbShim::new_in_memory()),
             SENDER_PRIV_KEY.public().to_address(),
         );