+++ packages/core/src/index.ts
@@ -337,7 +337,13 @@ export class Hopr extends EventEmitter {
     let packetCfg = new PacketInteractionConfig(this.packetKeypair, this.chainKeypair)
     packetCfg.check_unrealized_balance = this.options.checkUnrealizedBalance ?? true
 
-    const onReceivedMessage = (msg: ApplicationData) => this.emit('hopr:message', msg)
+    const onReceivedMessage = (msg: Uint8Array) => {
+      try {
+        this.emit('hopr:message', ApplicationData.deserialize(msg))
+      } catch (err) {
+        log(`could not deserialize application data: ${err}`)
+      }
+    }
 
     log('Linking chain and packet keys')
     this.db.link_chain_and_packet_keys(this.chainKeypair.to_address(), this.packetKeypair.public(), Snapshot._default())
@@ -1413,7 +1419,7 @@ export class Hopr extends EventEmitter {
     return result
   }
 
-  private async peerIdToChainKey(id: PeerId) {
+  public async peerIdToChainKey(id: PeerId): Promise<Address> {
     let pk = OffchainPublicKey.from_peerid_str(id.toString())
     return await this.db.get_chain_key(pk)
   }
@@ -1423,13 +1429,16 @@ export class Hopr extends EventEmitter {
    * @returns true if allowed access
    */
   public async isAllowedAccessToNetwork(id: PeerId): Promise<boolean> {
-    let chain_key = await this.peerIdToChainKey(id)
-    if (chain_key) {
-      return HoprCoreEthereum.getInstance().isAllowedAccessToNetwork(Address.deserialize(chain_key.serialize()))
-    } else {
-      log(`failed to determine channel key of ${id.toString()}`)
-      return false
+    // Don't wait for key binding and local linking if identity is the local node
+    if (this.id.equals(id)) {
+      return await this.db.is_allowed_to_access_network(this.getEthereumAddress())
     }
+    let chain_key: Address = await this.peerIdToChainKey(id)
+    // Only check if we found a chain key, otherwise peer is not allowed
+    if (chain_key) {
+      return await this.db.is_allowed_to_access_network(chain_key)
+    }
+    return false
   }
 
   /**